Always call move-window-to-head in focus-window.
This fixes a focus bug affecting the Arduino IDE.  The IDE window is
not necessarily the result of (screen-focus (screen)), so when
switching to the window that is, (eq window cw) and the window list is
not touched.

I'm not sure whether the fact that the IDE is not set as the screen
focus window even when it is focused is a bug in itself, but at any
rate it seems reasonable that 'focused window' and 'head of window
list' are different concepts (since windows don't have to take focus
at all).